requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
frozenlist==1.2.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==5.2.0
platformdirs==2.4.1
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
toml==0.10.2
wrapt==1.13.3
yarl==1.7.2
